requirements/base.txt
Django>=5.1,<6.0
django-filter==25.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
djangorestframework>=3.15,<4.0
djangorestframework_simplejwt>=5.5,<6.0
